+++ client/src/test/java/org/realityforge/gwt/appcache/client/TestApplicationCache.java
@@ -27,4 +27,9 @@ final class TestApplicationCache
   public void update()
   {
   }
+
+  @Override
+  public void removeCache()
+  {
+  }
 }